# Allow parsing in skipped row for C engine