@gfyoung 2016-04-14
The issue with allowing parsing on skipped rows is that there is a performance hit since we're spending more time processing rows that don't really matter.  How significant is it though do you think?