@jreback 2016-04-16
this for sure needs performance testing.