@gfyoung 2016-04-16
@jreback : it's actually a little more general than that.  I am addressing all of the cases that occur in `tokenizer.c`, which goes beyond just newline characters in quoted fields bug.  The other option is to add new logic in the `SKIP_LINE` state to check for quotes but I wasn't sure about how it would handle quotes within quotes?