@gfyoung 2016-04-16
"special case thinks" ?