@gfyoung 2016-04-16
@jreback : Added some additional unit tests that my current fixes should be robust against.  Let me know if there are any additional unit tests that should be included/written.