@jorisvandenbossche 2016-04-17
@gfyoung There is one specific asv bench for skiprows (io_bench/read_csv_skiprows), but you can eg run all tests that contain `read_csv`, or all tests in the io_bench.py file