@gfyoung 2016-04-17
@jorisvandenbossche : how do I exactly run only one of the benchmark tests (i.e. just `io_bench.py`)?  The `vb_suite/run_suite.py` file doesn't seem to have any way to configure which benchmarks to run.