@gfyoung 2016-04-17
@jorisvandenbossche : Thank you!  That link was very helpful.  I was looking on the `github.io` page for `vbench`, which was not as detailed as this.

<b>@Everyone</b>: I ran the benchmark for `io_bench.csv_skiprows`.  `master` currently runs it in 18.62 ms, while my PR runs it in 26.76 ms.