@jreback 2016-04-17
pls run all of the csv benchmarks