@gfyoung 2016-04-17
|                                                                            | master (ms) | PR (ms) |
| ---------------------------------------------------------------|:----------------:| -----------:|
| time_frame_to_csv                                             | 220.30         | 227.45    |
| time_frame_to_csv2                                           | 377.33         | 364.85    |
| time_frame_to_csv_date_formatting                  | 17.01           | 15.51      |
| time_frame_to_csv_mixed                                  | 269.36        | 257.73    |
| time_read_csv_infer_datetime_format_custom  | 18.10          | 18.30      |
| time_read_csv_infer_datetime_format_iso8601 | 2.97            | 3.10        |
| time_read_csv_infer_datetime_format_ymd       | 3.30            | 3.21        |
| time_read_csv_skiprows                                     | 18.23          | 25.74      |
| time_read_csv_standard                                     | 14.95          | 15.65      |
| time_read_parse_dates_iso8601                        | 2.05            | 2.14        |
| time_write_csv_standard                                     | 58.79          | 59.01      |

As can be observed according to `asv`, the performance has not been significantly impacted by my changes AFAICT.