@gfyoung 2016-04-17
If there are no complaints about the performance or the changes (they pass currently against Travis, I added `[no skip]` to change commit message), this PR is ready to merge.