@gfyoung 2016-04-17
I took a look through all of the issues, and I think the two that I mentioned so far are the only ones that can be closed with this PR.