@gfyoung 2016-04-17
@jreback : I looked through all of the open issues tagged with `CSV` before I made that comment above.