@gfyoung 2016-04-17
@jreback : if there is nothing else, I think this is ready to merge.