@jorisvandenbossche 2016-04-17
For the skiprows benchmark, there is a considerable slowdown, which is of course logical since there is now more parsing, and the exact slowdown of course depends on the number of skipped rows (this is just the consequence of wanting to solve the bug). 
But the question is a bit of we are OK with this performance hit for such a corner case. I can imagine people use `skiprows` exactly to have a faster parsing when they need only part of the file? 
However, I also can't think of a nice way around this (apart from yet another keyword ..).