@gfyoung 2016-04-17
@jorisvandenbossche : Yes, that slowdown seems to be an unfortunate necessity at this point for the sake of correctness.  As I mentioned earlier, the other option was to add more logic under the `SKIP_LINE` block, but attempts to do so were not robust against additional testing and only appeared to duplicate existing code.