@gfyoung 2016-04-21
@jorisvandenbossche , @jreback : I'm going to table this one for now until #12939 gets merged in.  It will be a lot easier to merge a fix for this with only one (hopefully) tokenizing function to work with.

I also realized that one potential issue with my current fix is that we will be penalizing people if the fields in the skipped row are malformed, which is not what we want IINM.  Thus, I might need to go back to the drawing board (even if the correctness > the performance as the PR currently stands) to see whether I can build in some other cases and not penalize.