@gfyoung 2016-04-22
I knew there had to be a better way.  Just needed some fresh eyes on the issue and managed to get the adding of a couple of new states to work.  Let's see what Travis thinks.