@jreback 2016-04-22
ok pls rerun the perf tests as well (and maybe run vs 0.18.0) as a comparison