@gfyoung 2016-04-22
FYI: I think you have an old build from your most recent PR that you just merged running on Travis right now.